First Commit
finished all effort towards this simple scouting program in only 4 to 5
hours.  The new programming libraries make programming a new scouting
system a breeze, and this test went well.  Need to consider narrowing
information from the game onto one page when given lots and lots of
rules, because the more basic and fundamental things are, the better.